EPMLABSBRN-2299 [BE & FrontEnd] Refactoring: merge controllers StudyHistoryController and StudyHistoryControllerV2 into one class
