+++ src/main/kotlin/com/epam/brn/controller/StudyHistoryController.kt
@@ -1,8 +1,8 @@
 package com.epam.brn.controller
 
+import com.epam.brn.dto.StudyHistoryDto
 import com.epam.brn.dto.response.BaseResponse
 import com.epam.brn.dto.response.BaseSingleObjectResponse
-import com.epam.brn.dto.StudyHistoryDto
 import com.epam.brn.service.StudyHistoryService
 import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
@@ -15,6 +15,8 @@ import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.PathVariable
+import java.time.LocalDateTime
 
 @RestController
 @RequestMapping("/study-history")
@@ -40,4 +42,19 @@ class StudyHistoryController(@Autowired val studyHistoryService: StudyHistorySer
         @RequestParam("year", required = true) year: Int
     ) = ResponseEntity.ok()
         .body(BaseResponse(data = studyHistoryService.getMonthHistoriesForCurrentUser(month, year)))
+
+    @GetMapping("/histories")
+    @ApiOperation("Get current user's study histories for period from <= startTime < to. Where from and to are dates in ISO format")
+    fun getHistories(
+        @RequestParam("from", required = true) from: LocalDateTime,
+        @RequestParam("to", required = true) to: LocalDateTime
+    ) = ResponseEntity.ok()
+        .body(BaseResponse(data = studyHistoryService.getHistoriesForCurrentUser(from, to)))
+
+    @GetMapping("/user/{userId}/has/statistics")
+    @ApiOperation("Check if user has statistics")
+    fun isUserHasStatistics(
+        @PathVariable("userId") userId: Long
+    ) = ResponseEntity.ok()
+        .body(BaseSingleObjectResponse(data = studyHistoryService.isUserHasStatistics(userId)))
 }
