+++ src/test/kotlin/com/epam/brn/integration/StudyHistoryControllerIT.kt
@@ -1,8 +1,10 @@
 package com.epam.brn.integration
 
-import com.epam.brn.dto.response.BaseSingleObjectResponse
 import com.epam.brn.dto.StudyHistoryDto
+import com.epam.brn.dto.response.BaseResponse
+import com.epam.brn.dto.response.BaseSingleObjectResponse
 import com.epam.brn.repo.StudyHistoryRepository
+import com.fasterxml.jackson.core.type.TypeReference
 import com.google.gson.Gson
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Test
@@ -15,7 +17,9 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 import java.nio.charset.StandardCharsets
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
+import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
+import kotlin.test.assertTrue
 
 @WithMockUser(username = "test@test.test", roles = ["ADMIN"])
 class StudyHistoryControllerIT : BaseIT() {
@@ -81,5 +85,127 @@ class StudyHistoryControllerIT : BaseIT() {
 
         // THEN
         assertNotNull(singleObjectResponseDto)
+    }
+
+    @Test
+    @WithMockUser(username = "test@test.test", roles = ["ADMIN"])
+    fun `getHistories should return histories for period of time`() {
+        // GIVEN
+        val user = insertDefaultUser()
+        val exercise = insertDefaultExercise()
+        val exercisingYear = 2019
+        val exercisingMonth = 3
+        val studyHistoryFirst =
+            insertDefaultStudyHistory(user, exercise, LocalDateTime.of(exercisingYear, exercisingMonth, 20, 13, 0), 25)
+        val studyHistorySecond =
+            insertDefaultStudyHistory(user, exercise, LocalDateTime.of(exercisingYear, exercisingMonth, 20, 14, 0), 25)
+        val from = LocalDateTime.of(exercisingYear, exercisingMonth, 1, 1, 1)
+        val to = LocalDateTime.of(exercisingYear, exercisingMonth, 28, 1, 1)
+        val datePattern = DateTimeFormatter.ISO_DATE_TIME
+        val expectedStudyHistories = listOf(studyHistoryFirst.toDto(), studyHistorySecond.toDto())
+
+        // WHEN
+        val response = mockMvc.perform(
+            get("$baseUrl/histories")
+                .param(fromParameterName, from.format(datePattern))
+                .param(toParameterName, to.format(datePattern))
+        )
+            .andExpect(status().isOk)
+            .andReturn().response.getContentAsString(StandardCharsets.UTF_8)
+
+        // THEN
+        val data = objectMapper.readValue(response, BaseResponse::class.java).data
+        val studyHistories: List<StudyHistoryDto> =
+            objectMapper.readValue(
+                objectMapper.writeValueAsString(data),
+                object : TypeReference<List<StudyHistoryDto>>() {}
+            )
+
+        assertNotNull(studyHistories)
+        assertEquals(expectedStudyHistories, studyHistories)
+    }
+
+    @Test
+    @WithMockUser(username = "test@test.test", roles = ["USER"])
+    fun `getHistories should return histories for period of time for user with role user`() {
+        // GIVEN
+        val user = insertDefaultUser()
+        val exercise = insertDefaultExercise()
+        val exercisingYear = 2019
+        val exercisingMonth = 3
+        val studyHistoryFirst =
+            insertDefaultStudyHistory(user, exercise, LocalDateTime.of(exercisingYear, exercisingMonth, 20, 13, 0), 25)
+        val studyHistorySecond =
+            insertDefaultStudyHistory(user, exercise, LocalDateTime.of(exercisingYear, exercisingMonth, 20, 14, 0), 25)
+        val from = LocalDateTime.of(exercisingYear, exercisingMonth, 1, 1, 1)
+        val to = LocalDateTime.of(exercisingYear, exercisingMonth, 28, 1, 1)
+        val datePattern = DateTimeFormatter.ISO_DATE_TIME
+        val expectedStudyHistories = listOf(studyHistoryFirst.toDto(), studyHistorySecond.toDto())
+
+        // WHEN
+        val response = mockMvc.perform(
+            get("$baseUrl/histories")
+                .param(fromParameterName, from.format(datePattern))
+                .param(toParameterName, to.format(datePattern))
+        )
+            .andExpect(status().isOk)
+            .andReturn().response.getContentAsString(StandardCharsets.UTF_8)
+
+        // THEN
+        val data = objectMapper.readValue(response, BaseResponse::class.java).data
+        val studyHistories: List<StudyHistoryDto> =
+            objectMapper.readValue(
+                objectMapper.writeValueAsString(data),
+                object : TypeReference<List<StudyHistoryDto>>() {}
+            )
+
+        assertNotNull(studyHistories)
+        assertEquals(expectedStudyHistories, studyHistories)
+    }
+
+    @Test
+    @WithMockUser(username = "test@test.test", roles = ["ADMIN"])
+    fun `isUserHasStatistics should return true when user has statistics`() {
+        // GIVEN
+        val user = insertDefaultUser()
+        val exercise = insertDefaultExercise()
+        val exercisingYear = 2019
+        val exercisingMonth = 3
+        insertDefaultStudyHistory(user, exercise, LocalDateTime.of(exercisingYear, exercisingMonth, 20, 13, 0), 25)
+
+        // WHEN
+        val response = mockMvc.perform(
+            get("$baseUrl/user/{userId}/has/statistics", user.id)
+        ).andExpect(status().isOk)
+            .andReturn().response.getContentAsString(StandardCharsets.UTF_8)
+
+        // THEN
+        val data = objectMapper.readValue(response, BaseSingleObjectResponse::class.java).data
+        val isUserHasStatistics: Boolean =
+            objectMapper.readValue(objectMapper.writeValueAsString(data), object : TypeReference<Boolean>() {})
+        assertTrue(isUserHasStatistics)
+    }
+
+    @Test
+    @WithMockUser(username = "test@test.test", roles = ["USER"])
+    fun `isUserHasStatistics should return true when user has statistics for user with role user`() {
+        // GIVEN
+        val user = insertDefaultUser()
+        val exercise = insertDefaultExercise()
+        val exercisingYear = 2019
+        val exercisingMonth = 3
+        insertDefaultStudyHistory(user, exercise, LocalDateTime.of(exercisingYear, exercisingMonth, 20, 13, 0), 25)
+
+        // WHEN
+        val response = mockMvc.perform(
+            get("$baseUrl/user/{userId}/has/statistics", user.id)
+        ).andExpect(status().isOk)
+            .andReturn().response.getContentAsString(StandardCharsets.UTF_8)
+
+        // THEN
+        val data = objectMapper.readValue(response, BaseSingleObjectResponse::class.java).data
+        val isUserHasStatistics: Boolean =
+            objectMapper.readValue(objectMapper.writeValueAsString(data), object : TypeReference<Boolean>() {})
+        assertTrue(isUserHasStatistics)
     }
 }
